+++ spec/unit/resource/api1200/synergy/volume_spec.rb
@@ -11,10 +11,10 @@
 
 require 'spec_helper'
 
-RSpec.describe OneviewSDK::API1000::Synergy::StorageSystem do
+RSpec.describe OneviewSDK::API1200::Synergy::Volume do
   include_context 'shared context'
 
-  it 'inherits from OneviewSDK::API1000::C7000::StorageSystem' do
-    expect(described_class).to be < OneviewSDK::API1000::C7000::StorageSystem
+  it 'inherits from OneviewSDK::API1200::C7000::Volume' do
+    expect(described_class).to be < OneviewSDK::API1200::C7000::Volume
   end
 end