+++ spec/unit/resource/api1600/c7000/logical_interconnect_group_spec.rb
@@ -11,10 +11,17 @@
 
 require 'spec_helper'
 
-RSpec.describe OneviewSDK::API1000::Synergy::StorageSystem do
+RSpec.describe OneviewSDK::API1600::C7000::LogicalInterconnectGroup do
   include_context 'shared context'
 
-  it 'inherits from OneviewSDK::API1000::C7000::StorageSystem' do
-    expect(described_class).to be < OneviewSDK::API1000::C7000::StorageSystem
+  it 'inherits from OneviewSDK::API1200::C7000::LogicalInterconnectGroup' do
+    expect(described_class).to be < OneviewSDK::API1200::C7000::LogicalInterconnectGroup
+  end
+
+  describe '#initialize' do
+    it 'should be initialize the instance with default values' do
+      item = described_class.new(@client_1600)
+      expect(item['type']).to eq('logical-interconnect-groupV8')
+    end
   end
 end