+++ spec/unit/resource/api800/c7000/storage_pool_spec.rb
@@ -11,10 +11,10 @@
 
 require 'spec_helper'
 
-RSpec.describe OneviewSDK::API1000::Synergy::StorageSystem do
+RSpec.describe OneviewSDK::API800::C7000::StoragePool do
   include_context 'shared context'
 
-  it 'inherits from OneviewSDK::API1000::C7000::StorageSystem' do
-    expect(described_class).to be < OneviewSDK::API1000::C7000::StorageSystem
+  it 'inherits from OneviewSDK::API600::C7000::StoragePool' do
+    expect(described_class).to be < OneviewSDK::API600::C7000::StoragePool
   end
 end